refactor api following (#218)
