+++ api/index.js
@@ -1,10 +1,9 @@
 const _ = require('lodash')
 const { API_DEADLINE, API_HOST, API_PORT, API_PROTOCOL, API_TIMEOUT, } = require('./config')
 const { GCP_FILE_BUCKET, GOOGLE_RECAPTCHA_SECRET, GCS_IMG_MEMBER_PATH, GCS_IMG_POST_PATH, } = require('./config')
-const { ENDPOINT_SECURE, } = require('./config')
 const { SERVER_PROTOCOL, SERVER_HOST, SERVER_PORT, } = require('./config')
 const { camelizeKeys, } = require('humps')
-const { constructScope, fetchPermissions, } = require('./services/perm')
+const { authorize, constructScope, fetchPermissions, } = require('./services/perm')
 const { initBucket, makeFilePublic, uploadFileToBucket, deleteFilesInFolder, publishAction, } = require('./gcs.js')
 const { processImage, } = require('./sharp.js')
 const { verifyToken, } = require('./middle/member/comm')
@@ -113,29 +112,6 @@ const fetchPromise = (url) => {
   })
 }
 
-const authorize = (req, res, next) => {
-  const whitelist = _.get(ENDPOINT_SECURE, [ `${req.method}${req.url.replace(/\?[A-Za-z0-9.*+?^=!:${}()#%~&_@\-`|[\]/\\]*$/, '')}`, ])
-  if (whitelist) {
-    fetchPermissions().then((perms) => {
-      Promise.all([
-        new Promise((resolve) => (resolve(_.get(whitelist, [ 'role', ]) ? _.find(_.get(whitelist, [ 'role', ]), (r) => (r === req.user.role)) : true))),
-        new Promise((resolve) => (resolve(_.get(whitelist, [ 'perm', ]) ? _.get(whitelist, [ 'perm', ]).length === _.filter(_.get(whitelist, [ 'perm', ]), (p) => (_.find(_.filter(perms, { role: req.user.role, }), { object: p, }))).length : true))),
-      ]).then((isAuthorized) => {
-        const isRoleAuthorized = isAuthorized[ 0 ]
-        const isPermsAuthorized = isAuthorized[ 1 ]
-        if (isRoleAuthorized && isPermsAuthorized) {
-          next()
-        } else {
-          res.status(403).send('Forbidden. No right to access.').end()
-        }
-      })
-    })
-
-  } else {
-    next()
-  }
-}
-
 /**
  * 
  * special request handler
@@ -145,7 +121,6 @@ const authorize = (req, res, next) => {
 router.use('/activate', verifyToken, require('./middle/member/activation'))
 router.use('/following', require('./middle/following'))
 router.use('/initmember', authVerify, require('./middle/member/initMember'))
-router.use('/member/public', require('./middle/member'))
 router.use('/member', [ authVerify, authorize, ], require('./middle/member'))
 router.use('/comment', require('./middle/comment'))
 router.use('/register', authVerify, require('./middle/member/register'))
@@ -170,9 +145,6 @@ router.all('/post', [ authVerify, authorize, ], function(req, res, next) {
   next()
 })
 router.all('/posts', [ authVerify, authorize, ], function(req, res, next) {
-  next()
-})
-router.all('/following', [ authVerify, authorize, ], function(req, res, next) {
   next()
 })
 router.all('/tags', [ authVerify, authorize, ], function(req, res, next) {
