+++ api/middle/member/index.js
@@ -1,6 +1,5 @@
 const { constructScope, fetchPermissions, } = require('../../services/perm')
-const { redisFetching, redisWriting, } = require('../redisHandler')
-const { find, pick, } = require('lodash')
+const { find, } = require('lodash')
 const config = require('../../config')
 const debug = require('debug')('READR:api:member')
 const express = require('express')
@@ -101,42 +100,6 @@ router.delete('*', (req, res, next) => {
   const userid = req.url.split('/')[ 1 ]
   userid && banUserForTalk(userid).then(() => {
     next()
-  })
-})
-
-router.get('/profile/:id', (req, res) => {
-  const id = req.params.id
-  const url = `${apiHost}/member/${id}`
-  if (!id) { res.status(403).send(`Forbidden.`) }
-  debug('Abt to fetch public member data.')
-  debug('>>>', req.url)
-  // res.send('ok')
-  redisFetching(`member${req.url}`, ({ err, data, }) => {
-    if (!err && data) {
-      debug('Fetch public member data from Redis.')
-      debug('>>>', req.url)
-      const mem = JSON.parse(data)
-      res.json({
-        'items': mem['_items'].map((object) => pick(object, [ 'id', 'nickname', 'description', 'profile_image', ])),
-      })
-    } else {
-      superagent
-      .get(url)
-      .end((e, response) => {
-        debug('Fetchpublic member data from api.', url)
-        if (!e && response) {
-          redisWriting(`member${req.url}`, response.text)
-          const mem = JSON.parse(response.text)
-          res.json({
-            'items': mem['_items'].map((object) => pick(object, [ 'id', 'nickname', 'description', 'profile_image', ])),
-          })
-        } else {
-          res.status(response.status).send('{\'error\':' + e + '}')
-          console.error(`error during fetch data from: member${req.url}`)
-          console.error(e)  
-        }
-      })
-    }
   })
 })
